Set local variables at the directory level for Emacs.
There was a bug in the file-local variable specification; a missing
colon after indent-tabs-mode. This meant that emacs always showed an
error message when loading any file. Since these file-local variables
are the same for all source files it makes sense to set them for all
in one place (affects all files opened in C-mode.) Unfortunately this
has meant touching every source file in the project.