Experimental TMX loader.
This replaces `load_map` with code to load from a TMX (Tiled) file, with the
following caveats

* The structure of the TMX map (number and names of layers, properties)
  needs to conform pretty exactly to what the maptiled.py script creates.
* All the tileset information is ignored, KQ gets the tileset from the
  tileset property of the map
* Very little/no error checking

The code loads the maps into memory and then sets them up just as the old
code did, so no changes to the drawing or logic were required.
The idea was just to test the feasibility of this. It seems very possible.
The next steps would be to look at the various data structures, e.g.
`s_entity` and make them more C++-like, also deciding which parts should
be serialised.
The map drawing code could become more flexible but this can be done when
converting to SDL2 or whatever.