Fix freed memory reference in music.
Extend the lifetime of the std::string backing a repeatedly used cstring
to the block in which it is used.


Music worked for me (TM) in classic r910, with minor modifications to
pulseaudio configuration, but was broken in the latest master.

After bisecting to find where it broke (why are there so many broken
intermediary commits? Repeat backporting is not fun T_T), I discovered the
culprit was commit c345d35b4e9544f1bdf7d02bd3e8f362a6900e66.

When changing the cstring logic to std::string, KMusic::play_music just took
c_str from a temporary string, which immediately went out of scope and was
freed. This freed pointer is then used repeatedly throughout the function.
This presumably happened to continue to point to the value that it was expected
to have, just enough to pass testing at the time.