Combat: Use selected item, rather than initial
The index of the item to use in combat was being calculated exactly once,
using the item the cursor started on, rather than the item that the
player selected.

This resulted in very strange behaviour, such as correctly showing the
select player interface when trying to use a light cure, but internally
attempting to use a cure poison, and failing silently.