Load attributes from the attribute value_list
When saving in store_stats, attributes are stored in a format like this:
<attributes>300,600,300,700,700,5500,10000,10000,0,5000,0,100,0</attributes>

When loading them in load_attributes, it's expecting a series of child
attributes.

This results in all attributes being set to 1 both when loading and when
starting a new game, which makes progress extremely difficult. Even spells
such as cure1 will only restore a single hitpoint with these nullified stats!